feat: Reques recover
	- Criando template genérico para as telas relacionas a autenticação
	- Tela de solicitação de recuperação de senha
	- LoadingContext para facilitar a chamada de loadings
