+++ package.json
@@ -18,6 +18,7 @@
     "cpf-cnpj-validator": "^1.0.3",
     "formik": "^2.2.9",
     "react": "^18.2.0",
+    "react-code-input": "^3.10.1",
     "react-dom": "^18.2.0",
     "react-modal": "^3.16.1",
     "react-router-dom": "^6.4.3",
