+++ src/index.jsx
@@ -1,6 +1,14 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
+import { AuthProvider } from './context/authContext'
+import { LoadingProvider } from './context/loadingContext'
 
 const root = ReactDOM.createRoot(document.getElementById('root'))
-root.render(<App />)
+root.render(
+    <AuthProvider>
+      <LoadingProvider>
+        <App />
+      </LoadingProvider>
+    </AuthProvider>
+)
