--- src/pages/recovery.jsx
@@ -1,3 +1,0 @@
-export const Recovery = () => {
-    return <h1>Recovery</h1>
-}
