+++ src/routes/index.jsx
@@ -1,13 +1,13 @@
 import { Routes, Route, BrowserRouter } from 'react-router-dom'
-import { Login } from '../pages/login'
-import { Register } from '../pages/register'
 import { StudentHome } from '../pages/home/studentHome'
 import { AdminHome } from '../pages/home/adminHome'
 import { GestorHome } from '../pages/home/gestorHome'
 import { CoordinatorHome } from '../pages/home/coordinatorHome'
 import { RequireAuth } from '../components/auth'
 import { MainRoleChoice } from '../pages/home/mainRoleChoice'
-import { Recovery } from '../pages/recovery'
+import { RequestRecovery } from '../pages/auth/recovery/requestRecovery'
+import { Login } from '../pages/auth/login'
+import { Register } from '../pages/auth/register'
 
 export const AppRoutes = () => {
   return (
@@ -16,7 +16,9 @@ export const AppRoutes = () => {
         <Route path="/" element={<Login />} />
         <Route path="/login" element={<Login />} />
         <Route path="/register" element={<Register />} />
-        <Route path="/recovery" element={<Recovery />} />
+        <Route path="/request-recovery" element={<RequestRecovery />} />
+        <Route path="/validate-recovery" element={<RequestRecovery />} />
+        <Route path="/reset-password" element={<RequestRecovery />} />
         <Route path="/role-choice" element={<MainRoleChoice />} />
         <Route
           path="/admin"
