+++ src/style.css
@@ -1,13 +1,13 @@
 @import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;500;700&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600&display=swap');
 
-*{  
+*{
     margin: 0;
     padding: 0;
     font-family: 'Montserrat';
 }
 
-* i { 
+* i {
     cursor:pointer;
 }
