+++ src/utils/requestErrorMessage.util.js
@@ -2,4 +2,6 @@ const messages = {
     USER_ALREADY_REGISTERED: 'Nome de usuário ou email já cadastrado'
 }
 
-export const parseErrorMessage = (message) => messages[message]
+export const parseErrorMessage = (message) => {
+    return messages[message] || 'Ocorreu um erro durante a solicitação'
+}
